project/requirements.txt
google-genai
# Project dependencies for hybrid chunking
tiktoken
sentence-transformers
chromadb
pdfplumber
scikit-learn
tqdm
numpy
nltk
